+++ src/EntityGraphQL/Schema/BaseField.cs
@@ -134,6 +134,63 @@ namespace EntityGraphQL.Schema
             ArgumentsAreInternal = true;
             UseArgumentsFromField = field;
         }
+        /// <summary>
+        /// To access this field all roles listed here are required
+        /// </summary>
+        /// <param name="roles"></param>
+        public IField RequiresAllRoles(params string[] roles)
+        {
+            if (RequiredAuthorization == null)
+                RequiredAuthorization = new RequiredAuthorization();
+            RequiredAuthorization.RequiresAllRoles(roles);
+            return this;
+        }
+
+        /// <summary>
+        /// To access this field any role listed is required
+        /// </summary>
+        /// <param name="roles"></param>
+        public IField RequiresAnyRole(params string[] roles)
+        {
+            if (RequiredAuthorization == null)
+                RequiredAuthorization = new RequiredAuthorization();
+            RequiredAuthorization.RequiresAllRoles(roles);
+            return this;
+        }
+
+        /// <summary>
+        /// To access this field all policies listed here are required
+        /// </summary>
+        /// <param name="policies"></param>
+        public IField RequiresAllPolicies(params string[] policies)
+        {
+            if (RequiredAuthorization == null)
+                RequiredAuthorization = new RequiredAuthorization();
+            RequiredAuthorization.RequiresAllPolicies(policies);
+            return this;
+        }
+
+        /// <summary>
+        /// To access this field any policy listed is required
+        /// </summary>
+        /// <param name="policies"></param>
+        public IField RequiresAnyPolicy(params string[] policies)
+        {
+            if (RequiredAuthorization == null)
+                RequiredAuthorization = new RequiredAuthorization();
+            RequiredAuthorization.RequiresAnyPolicy(policies);
+            return this;
+        }
+
+        /// <summary>
+        /// Clears any authorization requirements for this field
+        /// </summary>
+        /// <returns></returns>
+        public IField ClearAuthorization()
+        {
+            RequiredAuthorization = null;
+            return this;
+        }
 
         public IField AddValidator<TValidator>() where TValidator : IArgumentValidator
         {